Adding SVG icons for status labels
Removing duped/unused test expectations
De-nesting BEM rules
Remove unused comments
i18n
