+++ packages/edit-site/src/components/sidebar-navigation-data-list/style.scss
@@ -2,6 +2,6 @@
 	color: $gray-600;
 }
 
-.edit-site-sidebar-navigation_data-list__item .edit-site-sidebar-navigation_data-list__value {
+.edit-site-sidebar-navigation_data-list__value.edit-site-sidebar-navigation_data-list__value {
 	color: $white;
 }
