+++ packages/edit-site/src/components/sidebar-navigation-screen-page/style.scss
@@ -1,8 +1,6 @@
-.edit-site-sidebar-navigation-screen-page {
-	.edit-site-sidebar-navigation-screen__sticky-section {
-		padding: $grid-unit-40 0;
-		margin: $grid-unit-40 0;
-	}
+.edit-site-sidebar-navigation-screen__sticky-section {
+	padding: $grid-unit-40 0;
+	margin: $grid-unit-40 0;
 }
 
 .edit-site-sidebar-navigation-screen-page__featured-image-wrapper {
@@ -10,22 +8,23 @@
 	background-color: $gray-800;
 	margin-bottom: $grid-unit-20;
 	min-height: 128px;
-	.edit-site-sidebar-navigation-screen-page__featured-image {
-		border-radius: 2px;
-		height: 128px;
-		overflow: hidden;
-		width: 100%;
-		background-size: cover;
-		background-position: 50% 50%;
-		display: flex;
-		align-items: center;
-		justify-content: center;
-		color: $gray-600;
-	}
+}
 
-	.edit-site-sidebar-navigation-screen-page__featured-image-description {
-		font-size: $helptext-font-size;
-	}
+.edit-site-sidebar-navigation-screen-page__featured-image {
+	border-radius: 2px;
+	height: 128px;
+	overflow: hidden;
+	width: 100%;
+	background-size: cover;
+	background-position: 50% 50%;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	color: $gray-600;
+}
+
+.edit-site-sidebar-navigation-screen-page__featured-image-description {
+	font-size: $helptext-font-size;
 }
 
 .edit-site-sidebar-navigation-screen-page__excerpt {
@@ -41,70 +40,26 @@
 	}
 }
 
-.edit-site-sidebar-navigation-screen-page__status .components-text {
-	color: $white;
-}
+.edit-site-sidebar-navigation-screen-page__status {
+	display: inline-flex;
 
-.edit-site-sidebar-navigation-screen-page__status-indicator {
-	display: none;
-	&.has-status {
-		border-radius: 999px;
+	time {
+		display: contents;
+	}
+
+	svg {
 		height: 16px;
 		width: 16px;
-		display: flex;
-		align-items: center;
-		position: relative;
-		justify-content: center;
-
-		&::before {
-			height: 12px;
-			width: 12px;
-			border-radius: 60%;
-			border: 2px solid $gray-900;
-			display: block;
-			content: "";
-			position: absolute;
-		}
+		margin-right: $grid-unit-10;
 	}
 
-	&.has-publish-status {
-		background-color: $alert-green;
-		&::before {
-			background-color: $alert-green;
-			border: none;
-			width: inherit;
-			height: inherit;
-			content: "\2713";
-			text-align: center;
-			color: $gray-900;
-			font-weight: 700;
-		}
+	&.has-publish-status svg,
+	&.has-future-status svg {
+		fill: $alert-green;
 	}
 
-	&.has-future-status {
-		background-color: $alert-green;
-		&::before {
-			background-color: $alert-green;
-		}
-	}
-
-	&.has-pending-status {
-		background-color: $alert-yellow;
-		&::before {
-			background-color: $alert-yellow;
-		}
-	}
-
-	&.has-draft-status {
-		background-color: $alert-yellow;
-		&::before {
-			background-color: $gray-900;
-			border: 0;
-			border-radius: 0 0 8px 8px;
-			height: 8px;
-			left: -1px;
-			transform: rotate(90deg);
-			width: 13px;
-		}
+	&.has-pending-status svg,
+	&.has-draft-status svg {
+		fill: $alert-yellow;
 	}
 }
